Feat : Ephemeral Access Message

+++ src/main/java/fr/dawankama/o_drey/discord/listeners/Odrey.java
@@ -1,6 +1,5 @@
 package fr.dawankama.o_drey.discord.listeners;
 
-import fr.dawankama.o_drey.discord.events.DiscordEvent;
 import fr.dawankama.o_drey.discord.events.LoginEvent;
 import fr.dawankama.o_drey.discord.events.OtherEvent;
 import fr.dawankama.o_drey.discord.events.TjEvent;
@@ -14,34 +13,30 @@ import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.*;
-import net.dv8tion.jda.api.entities.channel.unions.MessageChannelUnion;
 import net.dv8tion.jda.api.events.guild.GuildReadyEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
-import net.dv8tion.jda.api.interactions.commands.build.CommandData;
-import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import net.dv8tion.jda.api.requests.GatewayIntent;
 import net.dv8tion.jda.api.utils.ChunkingFilter;
 import net.dv8tion.jda.api.utils.MemberCachePolicy;
 import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.event.EventListener;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 import org.springframework.stereotype.Service;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class Odrey extends ListenerAdapter {
@@ -51,10 +46,10 @@ public class Odrey extends ListenerAdapter {
     private final ApplicationEventPublisher publisher;
     private final ThreadPoolTaskScheduler scheduler;
     private Role banned;
-    public Odrey(SimpMessagingTemplate messagingTemplate, @Value("${discord.token}") String DISCORD_KEY, ApplicationEventPublisher publisher, ThreadPoolTaskScheduler scheduler) {
+    public Odrey(SimpMessagingTemplate messagingTemplate, @Value("${discord.token}") String key, ApplicationEventPublisher publisher, ThreadPoolTaskScheduler scheduler) {
         this.messagingTemplate = messagingTemplate;
         this.publisher = publisher;
-        this.jda = JDABuilder.createDefault(DISCORD_KEY)
+        this.jda = JDABuilder.createDefault(key)
                 .setMemberCachePolicy(MemberCachePolicy.ALL)
                 .enableIntents(GatewayIntent.GUILD_MEMBERS,
                         GatewayIntent.GUILD_MESSAGE_REACTIONS,
@@ -111,46 +106,57 @@ public class Odrey extends ListenerAdapter {
         Message message = event.getMessage();
         if(target != null && !target.getUser().isBot() && target.getRoles().stream().anyMatch(role -> role.getName().equals(banned.getName())))
             message.delete().queue();
-        if(message.getContentRaw().contains("n word")) {
+        if(message.getContentRaw().contains("Miaouh") && target != null) {
             message.reply("https://tenor.com/view/4k-caught-gif-20353888")
                     .queue(v -> event.getGuild().addRoleToMember(target, banned).queue(r -> scheduler.schedule(
                             () -> event.getGuild().removeRoleFromMember(target, banned).queue(f -> message.reply("https://tenor.com/view/out-of-jail-jail-prison-out-of-prison-im-out-gif-12660099").queue()
                             ),
-                            LocalDateTime.now().plusSeconds(30).atZone(ZoneId.systemDefault()).toInstant()
+                            LocalDateTime.now().plusMinutes(5).atZone(ZoneId.systemDefault()).toInstant()
                     )));
         }
     }
 
     @EventListener(LoginEvent.class)
     public void onLoginEvent(LoginEvent event) {
-        jda.getGuilds().stream()
-                .map(guild -> guild.getMembersByName(event.getUsername(), true))
-                .filter(list -> !list.isEmpty())
-                .findFirst()
-                .map(list -> list.get(0).getUser().openPrivateChannel())
-                .ifPresent(action -> action.queue(privateChannel ->
+        findUser(event.getUsername())
+                .ifPresent(user -> user.openPrivateChannel().queue(privateChannel ->
                         privateChannel.sendMessageEmbeds(new EmbedBuilder().setDescription("Tentative de connexion à votre compte sur DawAnkama").build())
                                 .setActionRow(
                                         Button.success("OK", "Valider"),
                                         Button.danger("KO", "Bloquer")
                                 )
                                 .queue(message -> scheduler.schedule(
-                                        () -> message.delete().queue(),
+                                        () -> {
+                                            try {
+                                                message.delete().queue();
+                                            } catch (Exception ignored) {/**/}
+                                        },
                                         LocalDateTime.now().plusMinutes(1).atZone(ZoneId.systemDefault()).toInstant()))));
+    }
+
+    @NotNull
+    public Optional<User> findUser(String name) {
+        return jda.getGuilds().stream()
+                .map(guild -> guild.getMembersByName(name, true))
+                .filter(list -> !list.isEmpty())
+                .findFirst()
+                .map(list -> list.get(0).getUser());
     }
 
     @Override
     public void onButtonInteraction(ButtonInteractionEvent event) {
         User user = event.getUser();
-        String destination = "/login/"+ user.getName();
-        event.reply("Décision prise en compte !").queue(r ->
-                {
-                    event.getMessage().delete().queue();
-                    messagingTemplate.convertAndSend(
-                            destination,
-                            new AuthResponse(user.getId(), user.getName(), "TOKEN")
-                    );
-                }
-        );
+        if (event.getButton().getId().equals("OK")) {
+            String destination = "/login/" + user.getName();
+            event.reply("Décision prise en compte !").setEphemeral(true).queue(r ->
+                    {
+                        event.getMessage().delete().queue();
+                        messagingTemplate.convertAndSend(
+                                destination,
+                                new AuthResponse(user.getId(), user.getName(), "TOKEN")
+                        );
+                    }
+            );
+        }
     }
 }
